pyproject.toml
[project]
name = "ai_doc"
version = "0.1.0"
description = ""
authors = [
    {name = "XRandolphX",email = "randolphfrp@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-cors (>=5.0.1,<6.0.0)",
    "openai (>=1.68.2,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "docxtpl (>=0.19.1,<0.20.0)",
    "pathlib (>=1.0.1,<2.0.0)",
]

[tool.poetry]
packages = [{include = "ai_doc", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
